ecommerce_perle/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0
